+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add experimental unused import analysis.
 - The option `--external-encodable-protocols` is deprecated, use `--external-codable-protocols` instead.
+- Assign-only properties on structs with synthesized initializers are now detected.
 
 ##### Bug Fixes
 
@@ -23,6 +24,7 @@
 - Fix public accessibility analysis false-positive for retained/ignored declarations.
 - Fix public accessibility analysis false-positive for enum case parameter types.
 - Fix public accessibility analysis false-positive for properties initialized with generic specialized types.
+- Types associated with assign-only properties are no longer identified as unused until the property is removed.
 
 ## 2.17.1 (2023-12-04)